support materials and rewrite a lot of stuff

+++ src/glsl/picking_fbo.frag
@@ -1,10 +1,10 @@
 #version 440 core
 
-out vec4 FragColor;
+out uvec4 FragColor;
   
 uniform uint objectIndex;
 
 void main()
 {
-    FragColor = vec4(objectIndex, 50, 20, 1);
+    FragColor = uvec4(objectIndex, 50, 20, 1);
 }
